+++ internal/query/user_schema.go
@@ -49,6 +49,10 @@ var (
 		name:  projection.UserSchemaIDCol,
 		table: userSchemaTable,
 	}
+	UserSchemaCreationDateCol = Column{
+		name:  projection.UserSchemaCreationDateCol,
+		table: userSchemaTable,
+	}
 	UserSchemaChangeDateCol = Column{
 		name:  projection.UserSchemaChangeDateCol,
 		table: userSchemaTable,
@@ -131,6 +135,7 @@ func NewUserSchemaStateSearchQuery(value domain.UserSchemaState) (SearchQuery, e
 func prepareUserSchemaQuery() (sq.SelectBuilder, func(*sql.Row) (*UserSchema, error)) {
 	return sq.Select(
 			UserSchemaIDCol.identifier(),
+			UserSchemaCreationDateCol.identifier(),
 			UserSchemaChangeDateCol.identifier(),
 			UserSchemaSequenceCol.identifier(),
 			UserSchemaInstanceIDCol.identifier(),
@@ -173,6 +178,7 @@ func prepareUserSchemaQuery() (sq.SelectBuilder, func(*sql.Row) (*UserSchema, er
 func prepareUserSchemasQuery() (sq.SelectBuilder, func(*sql.Rows) (*UserSchemas, error)) {
 	return sq.Select(
 			UserSchemaIDCol.identifier(),
+			UserSchemaCreationDateCol.identifier(),
 			UserSchemaChangeDateCol.identifier(),
 			UserSchemaSequenceCol.identifier(),
 			UserSchemaInstanceIDCol.identifier(),
