+++ proto/zitadel/object/v3alpha/object.proto
@@ -27,3 +27,11 @@ message Instance {
     string domain = 2;
   }
 }
+
+message Organization {
+  oneof property {
+    option (validate.required) = true;
+    string id = 1;
+    string domain = 2;
+  }
+}
