+++ proto/zitadel/resources/user/v3alpha/authenticator.proto
@@ -7,9 +7,9 @@ import "google/protobuf/struct.proto";
 import "google/protobuf/timestamp.proto";
 import "protoc-gen-openapiv2/options/annotations.proto";
 import "validate/validate.proto";
-import "zitadel/object/v2/object.proto";
+import "zitadel/resources/object/v3alpha/object.proto";
 
-option go_package = "github.com/zitadel/zitadel/pkg/grpc/user/v3alpha;user";
+option go_package = "github.com/zitadel/zitadel/pkg/grpc/resources/user/v3alpha;user";
 
 message Authenticators {
   // All of the user's usernames, which will be used for identification during authentication.
@@ -167,7 +167,7 @@ message AuthenticationKey {
       example: "\"69629023906488334\"";
     }
   ];
-  zitadel.object.v2.Details details = 2;
+  zitadel.resources.object.v3alpha.Details details = 2;
   // the file type of the key
   AuthNKeyType type = 3 [
     (grpc.gateway.protoc_gen_openapiv2.options.openapiv2_field) = {
