+++ proto/zitadel/resources/user/v3alpha/user.proto
@@ -7,22 +7,25 @@ import "google/protobuf/struct.proto";
 import "google/protobuf/timestamp.proto";
 import "protoc-gen-openapiv2/options/annotations.proto";
 import "validate/validate.proto";
-import "zitadel/object/v2/object.proto";
-import "zitadel/user/v3alpha/authenticator.proto";
-import "zitadel/user/v3alpha/communication.proto";
+import "zitadel/resources/object/v3alpha/object.proto";
+import "zitadel/resources/user/v3alpha/authenticator.proto";
+import "zitadel/resources/user/v3alpha/communication.proto";
 
-option go_package = "github.com/zitadel/zitadel/pkg/grpc/user/v3alpha;user";
+option go_package = "github.com/zitadel/zitadel/pkg/grpc/resources/user/v3alpha;user";
 
 message User {
-
-  // ID is the read-only unique identifier of the user.
-  string user_id = 1 [
+  // The user's data based on the provided schema.
+  google.protobuf.Struct data = 1 [
     (grpc.gateway.protoc_gen_openapiv2.options.openapiv2_field) = {
-      example: "\"69629012906488334\"";
+      example: "{\"name\":\"Gigi\",\"description\":\"the giraffe\"}"
     }
   ];
+}
+
+message GetUser{
   // Details provide some base information (such as the last change date) of the user.
-  zitadel.object.v2.Details details = 2;
+  zitadel.resources.object.v3alpha.Details details = 1;
+  User user = 2;
   // The user's authenticators. They are used to identify and authenticate the user
   // during the authentication process.
   Authenticators authenticators = 3;
@@ -31,9 +34,7 @@ message User {
   // State of the user.
   State state = 5;
   // The schema the user and it's data is based on.
-  Schema schema = 6;
-  // The user's data based on the provided schema.
-  google.protobuf.Struct data = 7;
+  GetSchema schema = 6;
 }
 
 enum State {
@@ -44,7 +45,7 @@ enum State {
   USER_STATE_LOCKED = 4;
 }
 
-message Schema {
+message GetSchema {
   // The unique identifier of the user schema.
   string id = 1 [
     (grpc.gateway.protoc_gen_openapiv2.options.openapiv2_field) = {
