+++ proto/zitadel/resources/userschema/v3alpha/user_schema.proto
@@ -1,25 +1,18 @@
 syntax = "proto3";
 
-package zitadel.user.schema.v3alpha;
+package zitadel.userschema.v3alpha;
 
 import "google/api/field_behavior.proto";
 import "google/protobuf/struct.proto";
 import "validate/validate.proto";
 import "protoc-gen-openapiv2/options/annotations.proto";
-import "zitadel/object/v2/object.proto";
+import "zitadel/resources/object/v3alpha/object.proto";
 
-option go_package = "github.com/zitadel/zitadel/pkg/grpc/user/schema/v3alpha";
+option go_package = "github.com/zitadel/zitadel/pkg/grpc/resources/userschema/v3alpha;userschema";
 
 message UserSchema {
-
-  // ID is the read-only unique identifier of the schema.
-  string id = 1 [
-    (grpc.gateway.protoc_gen_openapiv2.options.openapiv2_field) = {
-      example: "\"69629012906488334\""
-    }
-  ];
   // Details provide some base information (such as the last change date) of the schema.
-  zitadel.object.v2.Details details = 2;
+  zitadel.resources.object.v3alpha.Details details = 2;
   // Type is a human readable text describing the schema.
   string type = 3 [
     (grpc.gateway.protoc_gen_openapiv2.options.openapiv2_field) = {
@@ -60,55 +53,56 @@ enum FieldName {
   FIELD_NAME_STATE = 2;
   FIELD_NAME_REVISION = 3;
   FIELD_NAME_CHANGE_DATE = 4;
+  FIELD_NAME_CREATION_DATE = 5;
 }
 
-message SearchQuery {
-  oneof query {
+message SearchFilter {
+  oneof Filter {
     option (validate.required) = true;
 
-    // Union the results of each sub query ('OR').
-    OrQuery or_query = 1;
+    // Union the results of each sub filter ('OR').
+    OrFilter or_filter = 1;
     // Limit the result to match all sub queries ('AND').
     // Note that if you specify multiple queries, they will be implicitly used as andQueries.
-    // Use the andQuery in combination with orQuery and notQuery.
-    AndQuery and_query = 2;
-    // Exclude / Negate the result of the sub query ('NOT').
-    NotQuery not_query = 3;
+    // Use the andFilter in combination with orFilter and notFilter.
+    AndFilter and_filter = 2;
+    // Exclude / Negate the result of the sub filter ('NOT').
+    NotFilter not_filter = 3;
 
     // Limit the result to a specific schema type.
-    TypeQuery type_query = 5;
+    TypeFilter type_filter = 5;
     // Limit the result to a specific state of the schema.
-    StateQuery state_query = 6;
+    StateFilter state_filter = 6;
     // Limit the result to a specific schema ID.
-    IDQuery id_query = 7;
+    IDFilter id_filter = 7;
   }
 }
 
-message OrQuery {
-  repeated SearchQuery queries = 1 [
-    (grpc.gateway.protoc_gen_openapiv2.options.openapiv2_field) =  {
-      example: "[{\"idQuery\": {\"id\": \"163840776835432705\",\"method\": \"TEXT_QUERY_METHOD_EQUALS\"}},{\"idQuery\": {\"id\": \"163840776835943483\",\"method\": \"TEXT_QUERY_METHOD_EQUALS\"}}]"
+message OrFilter {
+  repeated SearchFilter queries = 1 [
+    (grpc.gateway.protoc_gen_openapiv2.options.openapiv2_field) = {
+      example: "[{\"idFilter\": {\"id\": \"163840776835432705\",\"method\": \"TEXT_FILTER_METHOD_EQUALS\"}},{\"idFilter\": {\"id\": \"163840776835943483\",\"method\": \"TEXT_FILTER_METHOD_EQUALS\"}}]"
     }
   ];
 }
-message AndQuery {
-  repeated SearchQuery queries = 1 [
-    (grpc.gateway.protoc_gen_openapiv2.options.openapiv2_field) =  {
-      example: "[{\"typeQuery\": {\"id\": \"employees\",\"method\": \"TEXT_QUERY_METHOD_STARTS_WITH\"}},{\"stateQuery\": {\"state\": \"STATE_ACTIVE\"}}]"
-    }
-  ];
-}
-
-message NotQuery {
-  SearchQuery query = 1 [
-    (grpc.gateway.protoc_gen_openapiv2.options.openapiv2_field) =  {
-      example: "{\"stateQuery\": {\"state\": \"STATE_ACTIVE\"}}"
+message AndFilter {
+  repeated SearchFilter queries = 1 [
+    (grpc.gateway.protoc_gen_openapiv2.options.openapiv2_field) = {
+      example: "[{\"typeFilter\": {\"id\": \"employees\",\"method\": \"TEXT_FILTER_METHOD_STARTS_WITH\"}},{\"stateFilter\": {\"state\": \"STATE_ACTIVE\"}}]"
     }
   ];
 }
 
-message IDQuery {
-  // Defines the ID of the user schema to query for.
+message NotFilter {
+  SearchFilter filter = 1 [
+    (grpc.gateway.protoc_gen_openapiv2.options.openapiv2_field) = {
+      example: "{\"stateFilter\": {\"state\": \"STATE_ACTIVE\"}}"
+    }
+  ];
+}
+
+message IDFilter {
+  // Defines the ID of the user schema to filter for.
   string id = 1 [
     (validate.rules).string = {min_len: 1, max_len: 200},
     (google.api.field_behavior) = REQUIRED,
@@ -118,14 +112,14 @@ message IDQuery {
       example: "\"163840776835432705\"";
     }
   ];
-  // Defines which text comparison method used for the id query.
-  zitadel.object.v2.TextQueryMethod method = 2 [
+  // Defines which text comparison method used for the id filter.
+  zitadel.resources.object.v3alpha.TextFilterMethod method = 2 [
     (validate.rules).enum.defined_only = true
   ];
 }
 
-message TypeQuery {
-  // Defines which type to query for.
+message TypeFilter {
+  // Defines which type to filter for.
   string type = 1 [
     (validate.rules).string = {max_len: 200},
     (google.api.field_behavior) = REQUIRED,
@@ -134,14 +128,14 @@ message TypeQuery {
       example: "\"employees\"";
     }
   ];
-  // Defines which text comparison method used for the type query.
-  zitadel.object.v2.TextQueryMethod method = 2 [
+  // Defines which text comparison method used for the type filter.
+  zitadel.resources.object.v3alpha.TextFilterMethod method = 2 [
     (validate.rules).enum.defined_only = true
   ];
 }
 
-message StateQuery {
-  // Defines the state to query for.
+message StateFilter {
+  // Defines the state to filter for.
   State state = 1 [
     (validate.rules).enum.defined_only = true,
     (google.api.field_behavior) = REQUIRED,
